*** Drive-A-Bot Code Changes *** This code runs the Drive-A-Bots and includes the controller enhancement with max motor speed set to 100.

+++ practice/darrell/controller_defines.h
@@ -1,8 +1,8 @@
 #ifndef CONTROLLER_DEFINES_H
 #define CONTROLLER_DEFINES_H
 
-const int ANALOG_DEAD_ZONE = 15;
-const int MAX_ANALOG = 128;
+const float ANALOG_DEAD_ZONE = 15.0;
+const float MAX_ANALOG = 128.0;
