+++ practice/darrell/motors.c
@@ -12,9 +12,9 @@
 // +MAX_ANALOG       +MAX_MOTOR_SPEED
 
 // Assuming a linear relationship between these values, we can solve for the slope //
-const int LINEAR_MOTOR_CONTROL_SLOPE = (MAX_MOTOR_SPEED-MIN_MOTOR_SPEED)/(MAX_ANALOG-ANALOG_DEAD_ZONE);
+const float LINEAR_MOTOR_CONTROL_SLOPE = (MAX_MOTOR_SPEED-MIN_MOTOR_SPEED)/(MAX_ANALOG-ANALOG_DEAD_ZONE);
 // Solve for the Y-intercept (equation was sovled for with a simple system of two equations assuming a linear relationship)
-const int LINEAR_MOTOR_CONTROL_INTERCEPT = (MIN_MOTOR_SPEED-(LINEAR_MOTOR_CONTROL_SLOPE*ANALOG_DEAD_ZONE));
+const float LINEAR_MOTOR_CONTROL_INTERCEPT = (MIN_MOTOR_SPEED-(LINEAR_MOTOR_CONTROL_SLOPE*ANALOG_DEAD_ZONE));
 
 // Calculates Output from Input based on a Linear Relationship
 int AttenuateControllerOutput(int output)
